move h2oModelOutput to dedicated h2oModelOutput directory
